Clean up codes
All the bindings in ops.lisp are defined using defmlx-method.
The methods and API are exposed in package.lisp.
The API should be considered as stable right now(?)

I will work on mlx-cl.image for my homework, so the API of
mlx-cl would not be modified for a while.

After finishing my homework, the API would be modified and it
should be like a numpy array manipulation and together with
test/op.lisp.

TODO:
+ mlx-cl.image (homework, have to be first priority)
+ test/op.lisp
+ lisp/comp-acc.lisp (for compiler macro to accelerate some operations)